20251022 backup
1. new project CSLibrary2024-NETStandard-TCP.csproj for CS203XL Network API mode
2. change API for multi antennas

+++ CSLibrary2024/Source/RFIDReader/CSLUnifiedAPI/Basic_API/ClassRFID.Public.Antenna.cs
@@ -391,12 +391,12 @@ namespace CSLibrary
             return Result.OK;
         }
 
-        public uint GetAntennaPort()
+        public int GetAntennaPort()
         {
             if (m_AntennaList == null)
                 return 1;
 
-            return (uint)m_AntennaList.Count;
+            return m_AntennaList.Count;
 		}
 
     }
